// src/xpanse-api/generated/models/Flavor.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* The flavors of the managed service
*/
export class Flavor {
    /**
    * The flavor name
    */
    'name': string;
    /**
    * The price of the flavor
    */
    'fixedPrice': number;
    /**
    * The properties of the flavor
    */
    'property': { [key: string]: string; };

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "fixedPrice",
            "baseName": "fixedPrice",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "property",
            "baseName": "property",
            "type": "{ [key: string]: string; }",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Flavor.attributeTypeMap;
    }

    public constructor() {
    }
}